requirements.txt
# Flask==2.3.3
# Werkzeug==2.3.7
# gunicorn==20.1.0
# python-dotenv==1.0.0

Flask==2.3.3
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
requests==2.31.0
twilio==8.9.0
gunicorn==20.1.0
sendgrid==6.10.0
python-dotenv==1.0.0